refactor: 로또 도메인 리팩토링, StringUtil GlobalUtil로 이름 변경
로또 도메인을 리팩토링 하며 검사 로직을 서비스에서 호출하도록 함, -
count의 경우 int로 반환하게 함, - 정렬 과정은 유틸리티에게 맡김

+++ src/main/java/lotto/repository/LottoRepository.java
@@ -1,6 +1,5 @@
 package lotto.repository;
 
-import java.math.BigInteger;
 import java.util.List;
 import lotto.model.Lotto;
 
@@ -8,7 +7,7 @@ public interface LottoRepository {
 
     void save(Lotto lotto);
 
-    BigInteger count();
+    int count();
 
     List<Lotto> findAll();
 }
