+++ src/main/java/lotto/repository/LottoRepositoryImpl.java
@@ -1,6 +1,5 @@
 package lotto.repository;
 
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import lotto.model.Lotto;
@@ -15,9 +14,8 @@ public class LottoRepositoryImpl implements LottoRepository {
     }
 
     @Override
-    public BigInteger count() {
-        //FIXME: storage.size의 반환 값이 int인 것에 대해
-        return BigInteger.valueOf(storage.size());
+    public int count() {
+        return storage.size();
     }
 
     @Override
